requirements.txt
mcp[cli]>=1.0.0
torch --extra-index-url https://download.pytorch.org/whl/cpu
diffusers>=0.33.0
transformers
accelerate
sentencepiece
protobuf
optimum-intel
openvino>=2025.1.0
nncf>=2.14.0
pillow
